Moved gesture threshold to physical units.

+++ src/client/src/NoteListGestureProcessor.h
@@ -21,7 +21,7 @@ private:
 private:
 
 	const double acceleration; // px/ms
-	const int    sensitivity;  // px
+	const double sensitivity;  // mm
 
 	IAnimator & animator;
